src/main/java/com/cyrex/awsimageuploader/profile/UserProfileDataAccessService.java
package com.cyrex.awsimageuploader.profile;

import com.cyrex.awsimageuploader.datastore.FakeUserProfileDataStore;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Repository;

import java.util.List;

@Repository
public class UserProfileDataAccessService {

    private final FakeUserProfileDataStore fakeUserProfileDataStore;

    @Autowired
    public UserProfileDataAccessService(FakeUserProfileDataStore fakeUserProfileDataStore) {
        this.fakeUserProfileDataStore = fakeUserProfileDataStore;
    }

    public List<UserProfile> getUserProfiles() {
        return fakeUserProfileDataStore.getUserProfiles();
    }

    public void addUser(UserProfile userProfile) {
        fakeUserProfileDataStore.addUser(userProfile);
    }

}
